Rename Mailing List Checkbox
Also updates description and examples for the field container

Ticket: https://phabricator.wikimedia.org/T394972
