--- src/pattern_library/css/blocks/mailing-list-checkbox.css
@@ -1,7 +1,0 @@
-.mailing-list-checkbox {
-	max-width: var(--form-field-max-width);
-}
-
-.mailing-list-checkbox > p {
-	padding-left: calc(0.5em + var(--space-s));
-}
