+++ src/pattern_library/css/styles.css
@@ -23,7 +23,7 @@
 @import "blocks/footer-nav.css";
 @import "blocks/icon-text.css";
 @import "blocks/locale.css";
-@import "blocks/mailing-list-checkbox.css";
+@import "blocks/verbose-checkbox.css";
 @import "blocks/mobile-nav-toggle.css";
 @import "blocks/modal-dialogue.css";
 @import "blocks/more-info-toggle.css";
