+++ src/pattern_library/patterns/field-container/description.md
@@ -1,1 +1,6 @@
-This is the container for a form field.
+This is the container for a form field. It handles the display of the fields, the help text and error states.
+It also contains a wrapper for when a grid of radio options is required.   
+
+## Usage
+
+To set the error state on a field you should add the attribute `data-error`.
