+++ src/pattern_library/patterns/field-container/markup.html
@@ -9,3 +9,30 @@
 	<input type="text" name="last-name" id="last-name" placeholder="It's a doggy dog world">
 	<p class="field-container__error-text">It's like a bowl in a china shop</p>
 </div>
+<fieldset class="field-container flow">
+	<legend>How often would you like to donate?</legend>
+	<div class="field-container__radio-grid">
+		<div class="grid" data-layout="halves">
+			<label>
+				<input type="radio" name="interval" id="interval-0">
+				<span>Single Payment</span>
+			</label>
+			<label>
+				<input type="radio" name="interval" id="interval-1">
+				<span>Monthly</span>
+			</label>
+			<label>
+				<input type="radio" name="interval" id="interval-3">
+				<span>Quarterly</span>
+			</label>
+			<label>
+				<input type="radio" name="interval" id="interval-6">
+				<span>Twice Yearly</span>
+			</label>
+			<label>
+				<input type="radio" name="interval" id="interval-12">
+				<span>Annually</span>
+			</label>
+		</div>
+	</div>
+</fieldset>
