+++ src/pattern_library/patterns/verbose-checkbox/description.md
@@ -1,0 +1,5 @@
+The verbose checkbox is a special checkbox in that it has a long info paragraph that follows it that needs to line up with the label text. It is used for the mailing list field.
+
+## Web Standards
+
+The input should have an `aria-describedby` attribute that points to the ID of the paragraph element.
