+++ src/pattern_library/patterns/verbose-checkbox/index.ts
@@ -1,10 +1,10 @@
 import { Pattern } from '@src/pattern_library/patterns/Pattern';
 
-const name = 'Mailing List Checkbox';
-const url = 'mailing-list-checkbox';
+const name = 'Verbose Checkbox';
+const url = 'verbose-checkbox';
 import description from './description.md';
 import html from './markup.html?raw';
 import examples from './Examples.vue';
-import css from '../../css/blocks/mailing-list-checkbox.css?raw';
+import css from '../../css/blocks/verbose-checkbox.css?raw';
 
 export default { name, url, description, html, examples, css } as Pattern;
